// web/src/generated/MessageDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Addr } from "./Addr";
import type { Attachment } from "./Attachment";
import type { MessageDates } from "./MessageDates";
import type { MessageFolder } from "./MessageFolder";

export type MessageDetail = { id: number, account_id: number, identity: string, msgid: string | null, subject: string, subject_norm: string, from: Array<Addr>, to: Array<Addr>, cc: Array<Addr>, dates: MessageDates, thread_id: string, thread_size: number, references: Array<string>, body_text: string, fresh_text: string, has_html: boolean, attachments: Array<Attachment>, folders: Array<MessageFolder>, images_allowed: boolean, };